+++ basecode_SDL2/CRanged.h
@@ -11,7 +11,7 @@ public:
 	CRanged(std::string type, std::string nom, int degats, float critique, int vie, float esquive, int vitesse, int attaque, int defense, int agilite, int intelligence);
 
 	virtual void Reparer();
-	virtual float GetSpecial();
+	virtual int GetSpecial();
 
 	virtual void print();
 };
