+++ basecode_SDL2/CStaff.h
@@ -13,7 +13,7 @@ public:
 	CStaff(std::string type, std::string nom, int degats, float critique, int vie, float esquive, int vitesse, int attaque, int defense, int agilite, int intelligence, int cout);
 
 	int GetCout();
-	float GetSpecial();
+	int GetSpecial();
 
 	void print();
 };
