+++ basecode_SDL2/CWeapon.cpp
@@ -100,9 +100,9 @@ int CWeapon::GetIntelligence()
 	return m_intelligence;
 }
 
-float CWeapon::GetSpecial()
+int CWeapon::GetSpecial()
 {
-	return 0.0f;
+	return 0;
 }
 
 float CWeapon::CalculerDegats()
